pyproject.toml
[project]
name = "rag-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "argon2-cffi>=23.1.0",
    "asgiref>=3.8.1",
    "autopep8>=2.3.2",
    "dj-database-url>=2.3.0",
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openai>=0.3.15",
    "llama-index-readers-file>=0.4.5",
    "llama-index-vector-stores-postgres>=0.4.2",
    "psycopg[binary]>=3.2.4",
    "pycodestyle>=2.12.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "sqlparse>=0.5.3",
    "toml>=0.10.2",
    "unipath>=1.1",
    "whitenoise>=6.9.0",
]
